Refactor script for domain resolution and dangling CNAME check
Removed the "config.json" file and the relevant sections of code that referenced it in "resolver.py". The "cname_checker.py" script was renamed to "dns_based_checks.py", optimizing the function to check for dangling CNAME records. Also updated the file "domain_processor.py" to reflect the name change. Generated a new zone file under the "example" directory.

not sure pattern matching is the way to go there.
