+++ imports/dns_based_checks.py
@@ -1,0 +1,44 @@
+"""
+This module provides methods to detect
+potential cloud service takeovers by identifying dangling CNAME records.
+
+Functions:
+--------
+detect_direct_takeovers(direct_reference_file, output_file):
+    Detect potential takeover candidates from a list of dangling CNAME records.
+
+check_dangling_cname(current_domain, nameservers):
+    Check if a domain has a dangling CNAME record.
+"""
+
+import dns.resolver
+
+
+def check_dangling_cname(current_domain, nameservers):
+    """
+    Check if a domain has a dangling CNAME record
+
+    :param current_domain: the domain to check
+    :type current_domain: str
+    :param nameservers: optional list of nameservers to use for resolution
+    :type nameservers: list[str]
+    :return: True if the domain has a dangling CNAME record, False otherwise
+    :rtype: bool
+    """
+    resolver = dns.resolver.Resolver()
+    if nameservers:
+        resolver.nameservers = nameservers
+
+    try:
+        resolver.resolve(current_domain, "A")
+        return False  # If resolving the A record succeeds, it's not a dangling CNAME
+    except (dns.resolver.NoAnswer, dns.resolver.NXDOMAIN, dns.resolver.NoNameservers):
+        try:
+            resolver.resolve(current_domain, "AAAA")
+            return False  # If resolving the AAAA record succeeds, it's not a dangling CNAME
+        except (
+            dns.resolver.NoAnswer,
+            dns.resolver.NXDOMAIN,
+            dns.resolver.NoNameservers,
+        ):
+            return True  # If both A and AAAA records cannot be resolved, it's a dangling CNAME
